Move external Dataset class into its own file
This is to be consistent with internal_dataset
